python/requirements.txt
PyMuPDF==1.23.14
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0